requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile pyproject.toml
#
aiohttp==3.9.3
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via gdown
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.13.1
    # via
    #   gdown
    #   huggingface-hub
    #   torch
fonttools==4.47.2
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.12.2
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
gdown==5.0.1
    # via ocr_tamil (pyproject.toml)
huggingface-hub==0.20.3
    # via timm
idna==3.6
    # via
    #   requests
    #   yarl
imageio==2.33.1
    # via scikit-image
jinja2==3.1.3
    # via torch
joblib==1.3.2
    # via nltk
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.3
    # via scikit-image
lightning-utilities==0.10.1
    # via
    #   pytorch-lightning
    #   torchmetrics
lmdb==1.4.1
    # via ocr_tamil (pyproject.toml)
markupsafe==2.1.4
    # via jinja2
matplotlib==3.8.2
    # via ocr_tamil (pyproject.toml)
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
nltk==3.8.1
    # via ocr_tamil (pyproject.toml)
numpy==1.26.3
    # via
    #   contourpy
    #   imageio
    #   matplotlib
    #   ocr_tamil (pyproject.toml)
    #   opencv-python
    #   pandas
    #   pytorch-lightning
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchmetrics
    #   torchvision
open-tamil==1.1
    # via ocr_tamil (pyproject.toml)
opencv-python==4.9.0.80
    # via ocr_tamil (pyproject.toml)
packaging==23.2
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   pytorch-lightning
    #   scikit-image
    #   torchmetrics
pandas==2.2.0
    # via ocr_tamil (pyproject.toml)
pillow==10.2.0
    # via
    #   imageio
    #   matplotlib
    #   ocr_tamil (pyproject.toml)
    #   scikit-image
    #   torchvision
pyparsing==3.1.1
    # via matplotlib
pysocks==1.7.1
    # via requests
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==1.9.5
    # via ocr_tamil (pyproject.toml)
pytz==2023.4
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   timm
regex==2023.12.25
    # via nltk
requests[socks]==2.31.0
    # via
    #   fsspec
    #   gdown
    #   huggingface-hub
    #   torchvision
safetensors==0.4.2
    # via timm
scikit-image==0.22.0
    # via ocr_tamil (pyproject.toml)
scipy==1.12.0
    # via scikit-image
six==1.16.0
    # via python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sympy==1.12
    # via torch
tifffile==2024.1.30
    # via scikit-image
timm==0.9.12
    # via ocr_tamil (pyproject.toml)
torch==2.1.2
    # via
    #   ocr_tamil (pyproject.toml)
    #   pytorch-lightning
    #   timm
    #   torchmetrics
    #   torchvision
torchmetrics==1.3.0.post0
    # via pytorch-lightning
torchvision==0.16.2
    # via
    #   ocr_tamil (pyproject.toml)
    #   timm
tqdm==4.66.1
    # via
    #   gdown
    #   huggingface-hub
    #   nltk
    #   pytorch-lightning
typing-extensions==4.9.0
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
tzdata==2023.4
    # via pandas
urllib3==2.2.0
    # via requests
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
